Use extension for bunker communication
Instead of signing/encrypting/decrypting the
events with a temporary random keypair, we can
use the nip07 extension (if installed).

This gives the advantage of not having to
re-approve a new app if you logged-out and
logged-in (suppose you don't want to keep
the session open in your browser). You can also
compare the npub of your extension with the npub that requests the permissions (in the nsec.app UI
or the nsecbunker.com UI).

Since NIP07 extensions don't have an api to get
the private key (and rightfully so!), I've changed
the session objects and the NostrConnectBroker
to use a connectSession object instead of a
connectKey - which has everything you need
to sign/encrypt/decrypt the communication
thanks to getSigner and getNip04 that were
already implemented.